Fix the test cases in daily run
