+++ cypress/e2e/bulk-edit/in-app/bulk-edit-in-app-items-matched-records-label.cy.js
@@ -1,0 +1,94 @@
+import permissions from '../../../support/dictionary/permissions';
+import BulkEditActions from '../../../support/fragments/bulk-edit/bulk-edit-actions';
+import BulkEditSearchPane from '../../../support/fragments/bulk-edit/bulk-edit-search-pane';
+import InventoryInstances from '../../../support/fragments/inventory/inventoryInstances';
+import TopMenu from '../../../support/fragments/topMenu';
+import Users from '../../../support/fragments/users/users';
+import FileManager from '../../../support/utils/fileManager';
+import getRandomPostfix from '../../../support/utils/stringTools';
+
+let user;
+const item = {
+  instanceName: `testBulkEdit_${getRandomPostfix()}`,
+  itemBarcode: getRandomPostfix(),
+};
+const invalidItemBarcodesFileName = `invalidItemBarcodes_${getRandomPostfix()}.csv`;
+const validItemBarcodesFileName = `validItemBarcodes_${getRandomPostfix()}.csv`;
+const validItemAccessionNumbersFileName = `validItemAccessionNumbers_${getRandomPostfix()}.csv`;
+const invalidBarcode = getRandomPostfix();
+
+describe('bulk-edit', () => {
+  describe('in-app approach', () => {
+    before('create test data', () => {
+      cy.createTempUser([
+        permissions.bulkEditView.gui,
+        permissions.bulkEditEdit.gui,
+        permissions.uiInventoryViewCreateEditItems.gui,
+      ]).then((userProperties) => {
+        user = userProperties;
+
+        InventoryInstances.createInstanceViaApi(item.instanceName, item.itemBarcode);
+
+        FileManager.createFile(
+          `cypress/fixtures/${invalidItemBarcodesFileName}`,
+          `${item.itemBarcode}\r\n${invalidBarcode}`,
+        );
+        FileManager.createFile(`cypress/fixtures/${validItemBarcodesFileName}`, item.itemBarcode);
+
+        cy.getItems({ limit: 1, expandAll: true, query: `"barcode"=="${item.itemBarcode}"` }).then(
+          (res) => {
+            const itemData = res;
+            itemData.accessionNumber = `testBulkEditAccessionNumber_${getRandomPostfix()}`;
+            cy.updateItemViaApi(itemData);
+            FileManager.createFile(
+              `cypress/fixtures/${validItemAccessionNumbersFileName}`,
+              itemData.accessionNumber,
+            );
+          },
+        );
+        cy.login(user.username, user.password, {
+          path: TopMenu.bulkEditPath,
+          waiter: BulkEditSearchPane.waitLoading,
+        });
+      });
+    });
+
+    after('delete test data', () => {
+      cy.getAdminToken();
+      InventoryInstances.deleteInstanceAndHoldingRecordAndAllItemsViaApi(item.itemBarcode);
+      Users.deleteViaApi(user.userId);
+      FileManager.deleteFile(`cypress/fixtures/${invalidItemBarcodesFileName}`);
+      FileManager.deleteFile(`cypress/fixtures/${validItemBarcodesFileName}`);
+      FileManager.deleteFile(`cypress/fixtures/${validItemAccessionNumbersFileName}`);
+    });
+
+    it(
+      'C357030 Verify Matched records label cleanup -- In -app approach (firebird)',
+      { tags: ['smoke', 'firebird', 'shiftLeft', 'C357030'] },
+      () => {
+        BulkEditSearchPane.checkItemsRadio();
+        BulkEditSearchPane.selectRecordIdentifier('Item barcode');
+
+        BulkEditSearchPane.uploadFile(invalidItemBarcodesFileName);
+        BulkEditSearchPane.waitFileUploading();
+
+        BulkEditSearchPane.verifyMatchedResults(item.itemBarcode);
+        BulkEditSearchPane.verifyNonMatchedResults(invalidBarcode);
+        BulkEditSearchPane.verifyErrorLabel(invalidItemBarcodesFileName, 1, 1);
+        BulkEditSearchPane.verifyPaneRecordsCount(1);
+
+        const newLocation = 'Annex';
+
+        BulkEditActions.openActions();
+        BulkEditActions.openInAppStartBulkEditFrom();
+        BulkEditActions.replaceTemporaryLocation(newLocation);
+        BulkEditActions.confirmChanges();
+        BulkEditActions.commitChanges();
+        BulkEditSearchPane.waitFileUploading();
+
+        BulkEditActions.verifySuccessBanner(1);
+        BulkEditSearchPane.verifyChangedResults(newLocation);
+      },
+    );
+  });
+});
