+++ cypress/support/fragments/bulk-edit/query-modal.js
@@ -42,7 +42,7 @@ export const usersFieldValues = {
   expirationDate: 'Users — User — Expiration date',
   firstName: 'Users — User — First name',
   lastName: 'Users — User — Last name',
-  patronGroup: 'Users — Group — Group',
+  patronGroup: 'Users — Group — Name',
   preferredContactType: 'Users — User — Preferred contact type',
   userActive: 'Users — User — Active',
   userBarcode: 'Users — User — Barcode',
